refactor: move AggregateRoot to another file
