+++ src/BankingApp.Domain.Core/IAggregateRoot.cs
@@ -1,5 +1,4 @@
 ﻿using MediatR;
-using System.Collections.ObjectModel;
 
 namespace BankingApp.Domain.Core;
 
@@ -12,19 +11,4 @@ public interface IAggregateRoot
     public void RemoveDomainEvent(INotification domainEvent);
 
     public void ClearDomainEvents();
-}
-
-public abstract class AggregateRoot<TId> : IEntity<TId>, IAggregateRoot
-{
-    private readonly List<INotification> _domainEvents = new();
-
-    public TId Id { get; protected set;  } = default!;
-
-    public IReadOnlyCollection<INotification> DomainEvents => new ReadOnlyCollection<INotification>(_domainEvents);
-
-    public void AddDomainEvent(INotification domainEvent) => _domainEvents.Add(domainEvent);
-
-    public void RemoveDomainEvent(INotification domainEvent) => _domainEvents.Remove(domainEvent);
-
-    public void ClearDomainEvents() => _domainEvents.Clear();
 }
